requirements.txt
torch
torchvision
numpy
Pillow
matplotlib
seaborn
scikit-learn
numpy
